apps/backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery>=5.3.0
redis>=4.0.0
gunicorn==21.2.0
Pillow>=10.0.0
